# Used SWC instead of Babel for Jest tests