--- a/package.json
+++ b/package.json
@@ -39,8 +39,7 @@
         "husky": "7.0.4",
         "jest": "27.5.1",
         "lint-staged": "12.3.4",
-        "prettier": "2.5.1",
-        "regenerator-runtime": "^0.13.9"
+        "prettier": "2.5.1"
     },
     "homepage": "https://github.com/typescript-eslint/tslint-to-eslint-config#readme",
     "keywords": [