--- a/package.json
+++ b/package.json
@@ -22,7 +22,6 @@
         "typescript": "4.5.5"
     },
     "devDependencies": {
-        "@swc/core": "^1.2.143",
         "@swc/jest": "^0.2.17",
         "@types/eslint-config-prettier": "6.11.0",
         "@types/glob": "7.2.0",