## PR Checklist

-   [x] Addresses an existing issue: fixes #1366
-   [x] That issue was marked as [`status: accepting prs`](https://github.com/typescript-eslint/tslint-to-eslint-config/labels/status%3A%20accepting%20prs)

## Overview

Uses `@swc/jest` instead of `babel-jest`.

-   Execution: `jest --clearCache` before each command. Terminal was in focus and no actions taken on open programs during test runs.
-   Hardware: Surface Laptop 4, 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz.
-   Reporting: Time ranges are rounded numbers from 7 runs with the highest and lowest values removed.
-   Software: WSL 2.0 in an Ubuntu drive via a VS Code terminal.

| Command                          | Babel Time    | SWC Time      | (old - new) / old x 100% |
| -------------------------------- | ------------- | ------------- | ------------------------ |
| `jest`                           | 18-20 seconds | 11-14 seconds | **30-40%** faster        |
| `jest --coverage --maxWorkers=2` | 23-28 seconds | 20-25 seconds | **10-13%** faster        |


Sorted raw time data:

* Babel:
    * `jest`: 16.151, 17.686, 18.548, 18.738, 19.278, 19.676, 20.091
    * `jest --coverage --maxWorkers=2`: 22.320, 23.239, 24.898, 25.109, 27.087, 27.724, 28.649
* SWC:
    * `jest`: 11.488, 11.862, 12.709, 13.558, 13.932, 14.215, 14.351
    * `jest --coverage --maxWorkers=2`: 19.839, 20.708, 22.430, 23.670, 24.704, 25.013, 25.248